feat: restructure authentication and profile components with new directory structure
