+++ next.config.mjs
@@ -2,6 +2,9 @@
 const nextConfig = {
   reactStrictMode: true,
   pageExtensions: ['page.tsx', 'page.ts'],
+  experimental: {
+    appDir: true,
+  },
 };
 
 export default nextConfig;
