+++ src/pages/auth/components/Footer.tsx
