+++ src/pages/auth/index.page.tsx
