+++ src/pages/profile/Profile.tsx
@@ -2,7 +2,7 @@ import ProfileName from './components/ProfileName';
 import ProfileImage from './components/ProfileImage';
 import ProfileContainer from './components/ProfileContainer';
 import Header from './components/Header';
-import CenterLayout from '@/src/components/CenterLayout';
+import CenterLayout from '@/components/CenterLayout';
 
 export default function Profile() {
   return (
